refactor(src): moved types to its own file and deleted useless files
