--- index.js
@@ -1,3 +1,0 @@
-import WOKCommands from "./src";
-
-export default WOKCommands;
