+++ src/commands/help/help.ts
@@ -1,6 +1,6 @@
 import { Client, Message, MessageEmbed } from "discord.js";
 import WOKCommands from "../..";
-import { ICallbackObject, ICommand } from "../../../typings";
+import { ICallbackObject, ICommand } from "../../types";
 import getFirstEmbed from "./!get-first-embed";
 import ReactionListener, { addReactions } from "./!ReactionListener";
