+++ src/commands/prefix.ts
@@ -1,5 +1,5 @@
-import { ICallbackObject, ICommand } from "../..";
 import prefixes from "../models/prefixes";
+import { ICallbackObject, ICommand } from "../types";
 
 export = {
   description: "Displays or sets the prefix for the current guild",
